backend/requirements.txt
streamlit
pdfplumber
azure-search-documents
azure-core
python-dotenv
langchain-openai